fix(ai-vault): split session refresh hook
